Tillat avviksrapportering på lukket liste
