+++ src/kandidat/kandidatliste/avviksrapportering/Avviksrapportering.module.css
@@ -1,0 +1,8 @@
+.gjennomgått {
+    display: flex;
+    gap: 0.5rem;
+}
+
+.gjennomgått svg {
+    font-size: 1.5rem;
+}
