+++ src/kandidat/kandidatliste/meny/Meny.module.css
@@ -8,18 +8,8 @@
     background-color: white;
 }
 
-.rapporterAvvik,
-.gjennomgått {
+.høyre {
     margin-left: auto;
-}
-
-.gjennomgått {
-    display: flex;
-    gap: 0.5rem;
-}
-
-.gjennomgått svg {
-    font-size: 1.5rem;
 }
 
 .border {
